tried a bunch of optimizations

+++ src/main/scala/BildScript/Transformations.scala
@@ -2,7 +2,7 @@ package BildScript
 
 object Transformations {
 
-  case class PositionTransform(x: Gen[Double], y: Gen[Double]) extends Transformation {
+  case class PositionTransform(x: Gen[Float], y: Gen[Float]) extends Transformation {
     override def next: Transformation = PositionTransform(x.nextGen, y.nextGen)
     override def exec(p: Point): Point = p - Point(x, y)
   }
